[Java][Improvement] Refactor project package
