+++ java/java.iml
@@ -7,6 +7,8 @@
     </content>
     <orderEntry type="inheritedJdk" />
     <orderEntry type="sourceFolder" forTests="false" />
-    <orderEntry type="library" exported="" name="javax.servlet-3.0" level="project" />
+    <orderEntry type="library" name="hamcrest-core-1.3" level="project" />
+    <orderEntry type="library" name="javax.servlet-3.0" level="project" />
+    <orderEntry type="library" name="junit-4.12" level="project" />
   </component>
 </module>
