--- java/src/Part2/Main.java
@@ -1,4 +1,0 @@
-package Part2;
-
-public class Main {
-}
