--- java/src/Part3/Main.java
@@ -1,4 +1,0 @@
-package Part3;
-
-public class Main {
-}
