--- java/src/Part4/Main.java
@@ -1,4 +1,0 @@
-package Part4;
-
-public class Main {
-}
