+++ java/src/com/framgia/Part1/bottle/BottleBetterVersion.java
@@ -1,12 +1,12 @@
-package Part3.bottle;
+package com.framgia.Part1.bottle;
 
-public class BottleStep1 {
+public class BottleBetterVersion {
 
     private static String song() {
         return getVerses(99, 0);
     }
 
-    private static String getVerses(int start, int end) {
+    private static String getVerses(int start, int end) { // GOOD: specific name
         StringBuilder poem = new StringBuilder();
         for (int i = start; i >= end; i--) {
             poem.append(getVerse(i));
@@ -15,7 +15,8 @@ public class BottleStep1 {
         return poem.toString();
     }
 
-    private static String getVerse(int number) {
+    private static String getVerse(int number) {  // GOOD: specific name
+        // GOOD: show the control flow and cases instead of hiding it in a big expression
         switch (number) {
             case 0:
                 return "No more bottles of beer on the wall, no more bottles of beer." + "\n" +
@@ -23,7 +24,7 @@ public class BottleStep1 {
             case 1:
                 return "1 bottle of beer on the wall, 1 bottle of beer." + "\n" +
                         "Take one down and pass it around, no more bottles of beer on the wall.";
-            case 2: // Can you identify the sub problem here and merge this case with the else case
+            case 2:
                 return "2 bottles of beer on the wall, 2 bottles of beer." + "\n" +
                         "Take one down and pass it around, 1 bottle of beer on the wall.";
             default:
