+++ java/src/com/framgia/Part1/getuserinfo/GetUserInfo.java
@@ -1,9 +1,9 @@
-package Part1.getuserinfo;
+package com.framgia.Part1.getuserinfo;
 
 import java.util.Calendar;
 import java.util.List;
 
-public class GetUserInfoWithComment {
+public class GetUserInfo {
 
     static class User {
         String birthday;
@@ -19,22 +19,15 @@ public class GetUserInfoWithComment {
         }
     }
 
-    // Can you come up with a better name here
-    // Find a specific alternative to get that perfectly describe this function
     private List<User> getUsers() {
-        // What is more important here?
-        // the format of the date
-        // or the date it self
-        // Come up with a better name for this variables.
         long yyyymmdstr = System.currentTimeMillis();
-        return RestClient.execute("GET", "http://example.com/users&starts_date=" + yyyymmdstr, 69); // what's is 69?
+        return RestClient.execute("GET", "http://example.com/users&starts_date=" + yyyymmdstr, 69);
     }
 
-    // Should we name this function like a getter?
     private int totalAge() {
         List<User> users = getUsers();
         int total = 0;
-        for (User u : users) { // Find a better name for this variable
+        for (User u : users) {
             int userBirthDay = getYearInTimeStr(u.birthday);
             int currentYear = Calendar.getInstance().get(Calendar.YEAR);
             total += currentYear - userBirthDay;
