+++ java/src/com/framgia/Part3/bottle/Bottle.java
@@ -1,6 +1,6 @@
-package Part3.bottle;
+package com.framgia.Part3.bottle;
 
-public class BottleStep2 {
+public class Bottle {
 
     private static String song() {
         return getVerses(99, 0);
@@ -16,8 +16,6 @@ public class BottleStep2 {
     }
 
     private static String getVerse(int number) {
-        // Now identify all the sub problems and merge all cases
-        // into one, remember that all sub problem should also only do one thing at a time
         switch (number) {
             case 0:
                 return "No more bottles of beer on the wall, no more bottles of beer." + "\n" +
@@ -30,15 +28,8 @@ public class BottleStep2 {
                         "Take one down and pass it around, 1 bottle of beer on the wall.";
             default:
                 return number + " bottles of beer on the wall, " + number + " bottles of beer." + "\n" +
-                        "Take one down and pass it around, " + getContainer(number - 1) + " of beer on the wall.";
+                        "Take one down and pass it around, " + (number - 1) + " bottles of beer on the wall.";
         }
-    }
-
-    /**
-     * The sub problem choosing the correct container name  base on the number of bottle
-     */
-    private static String getContainer(int number) {
-        return number == 1 ? "bottle" : "bottles";
     }
 
     public static void main(String[] args) {
